# Debug cache start pos